requirements.txt
python-telegram-bot==20.7
yfinance==0.2.35
APScheduler==3.10.4
python-dotenv==1.0.0 